+++ NEWS.md
@@ -1,5 +1,8 @@
 # inlmisc 0.4.4.9000
 
+- Add **alphahull** (released under a restrictive non-free software license)
+  and **maptools** to suggested packages.
+
 - Remove **httr** from suggested packages.
 
 - In `GetRegionOfInterest` function: add `alpha` and `width` arguments,